rename: 이벤트 모달 폴더 구성
